Update : Added a readme to the CPP module 03 for clear route planning.
The readme, is actually to explain my thought process so that i
can document where i am starting the project how to do so and
what are the things that must be known before starting it.

I also added the files that are nessersary for the project and
completed the makefile too.